Fixed a bug that caused quick updates to notion documents to be missed
Also added a feature that removes notion entries that have been deleted in the page. Before it just left them dangling.

Also fixed a small bug in the code for creating a new current day page. Two spaces were being inserted which caused the regex to not recognize the title as a date.